tcplink: Reap zombie execserver instances.
The docs for `exec.Command` state that the process doesn't get released
until `Wait()` is called on it. If we don't do this, the zombies build
up and eventually prevent us or anyone from creating more processes.

Use a goroutine so the current goroutine doesn't block; the reaping
happens while we're idle.