Support specifying environment variables to test processes.
This requires support from `execserver` to work, which means you need
a patch to the CTS itself.

The earlier version of Angular didn't really support separating lists
with whitespace. As a result, once saved, the environment variables
appeared in the `textarea` as a comma-separated list, even though the
`ng-list` directive set the separator to `'/\n/'`. This is probably
because the separator is given as a regex instead of a literal string.
But if it had a numeric literal `'&#10;'` instead of a regex, the list
_appeared_ correct in the UI... but the list in the database was
actually passed as a single large string with embedded newlines. For
this reason, the final version of AngularJS 1.3.0 is required.