Split the list of test cases into pages.
Fetch them one at a time.

Chrome struggles with large text messages, like the complete list of all
known test cases, sent over a WebSocket. By splitting it up into
multiple smaller messages, we can get around Chrome bug 1511182.

The one drawback is that this breaks the test suite, but we can't revert
this until the bug in Chrome is fixed. Unfortuantely, I don't know how
to make the test suite pass with this in place.

Add some more logging. This helped a bit with debugging this issue.