Update Angular to 1.3.0 final.
Update jQuery to 2.1.1. Update the custom tree control for changes in
Angular. Update ui-router to 0.2.12.

Conservatism is deliberate here, in order to minimize teething problems
relating to upgrading JS libraries.

This is a rather large change, but these libraries needed to be updated
together, or bugs start manifesting on the client like the tree control
not showing up, or the links on the Results page being broken.